RD-2002 Extract shared DeploymentsView component to widgets/common
Contains the implementation for the shared parts of top-level and
drilled-down Deployments View widgets.
