+++ widgets/common/src/deploymentsView/common.ts
@@ -1,6 +1,7 @@
 import type { SemanticICONS } from 'semantic-ui-react';
 
 export const i18nPrefix = 'widgets.deploymentsView';
+export const i18nMessagesPrefix = `${i18nPrefix}.messages`;
 
 export const subenvironmentsIcon: SemanticICONS = 'object group';
 export const subservicesIcon: SemanticICONS = 'cube';
