+++ widgets/common/src/deploymentsView/index.ts
@@ -1,22 +1,24 @@
-import * as Table from './table';
 import * as DCommon from './common';
-import DetailsPane from './detailsPane';
-import type { Deployment, DeploymentsResponse } from './types';
+import type { DeploymentsResponse } from './types';
 import { SharedDeploymentsViewWidgetConfiguration, sharedConfiguration, sharedDefinition } from './configuration';
 import './styles.scss';
+import { DeploymentsView as DDeploymentsView } from './DeploymentsView';
 
 declare global {
-    namespace Stage.Common.DeploymentsView {
-        // NOTE: necessary rename to DCommon, since `Common` resolves to `Stage.Common`, not the `Common` import
-        export { Table, DCommon as Common, DetailsPane, sharedDefinition };
+    namespace Stage.Common {
+        // eslint-disable-next-line @typescript-eslint/no-namespace
+        namespace DeploymentsView {
+            // NOTE: necessary rename to DCommon, since `Common` resolves to `Stage.Common`, not the `Common` import
+            export { DCommon as Common, sharedDefinition, DDeploymentsView as DeploymentsView };
 
-        // NOTE: no-namespace rule does not detect that `export namespace` are in a `declare` context
-        /* eslint-disable @typescript-eslint/no-namespace */
-        export namespace Configuration {
-            export { SharedDeploymentsViewWidgetConfiguration, sharedConfiguration };
-        }
-        export namespace Types {
-            export { Deployment, DeploymentsResponse };
+            // NOTE: no-namespace rule does not detect that `export namespace` are in a `declare` context
+            /* eslint-disable @typescript-eslint/no-namespace */
+            export namespace Configuration {
+                export { SharedDeploymentsViewWidgetConfiguration, sharedConfiguration };
+            }
+            export namespace Types {
+                export { DeploymentsResponse };
+            }
         }
     }
 }
@@ -25,10 +27,9 @@ Stage.defineCommon({
     name: 'DeploymentsView',
     common: {
         sharedDefinition,
-        DetailsPane,
         Common: DCommon,
-        Table,
         Configuration: { sharedConfiguration },
+        DeploymentsView: DDeploymentsView,
         Types: {}
     }
 });
